+++ packages/protocol/src/groups/FabricGroups.ts
@@ -7,11 +7,12 @@ import type { Fabric } from "#fabric/Fabric.js";
 import { BasicMap, Bytes, InternalError, MatterFlowError, StorageContext } from "#general";
 import { GroupKeySet, KeySets, OperationalKeySet } from "#groups/KeySets.js";
 import { MessagingState } from "#groups/MessagingState.js";
-import { GroupId } from "#types";
+import { GroupId, MATTER_EPOCH_OFFSET_US } from "#types";
 import { Groups } from "./Groups.js";
 
 export const GROUP_SECURITY_INFO = Bytes.fromString("GroupKey v1.0");
 
+export const IPK_DEFAULT_EPOCH_START_TIME = MATTER_EPOCH_OFFSET_US;
 /**
  * Class that contains an operational view on the Group Keys for a fabric
  */
@@ -34,7 +35,7 @@ export class FabricGroups {
             groupKeySetId: 0,
             epochKey0: fabric.identityProtectionKey,
             operationalEpochKey0: fabric.operationalIdentityProtectionKey,
-            epochStartTime0: 0, // 0 is always ok, but only for the IPK key
+            epochStartTime0: IPK_DEFAULT_EPOCH_START_TIME, // with the Year 2000-offset that's basically 0 and always ok, but only for the IPK key
             groupSessionId0: null,
             epochKey1: null,
             operationalEpochKey1: null,