+++ packages/types/src/tlv/TlvNumber.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import {
+    Days,
     FLOAT32_MAX,
     FLOAT32_MIN,
     FLOAT64_MAX,
@@ -16,6 +17,8 @@ import {
     INT64_MIN,
     INT8_MAX,
     INT8_MIN,
+    ImplementationError,
+    Seconds,
     UINT16_MAX,
     UINT24_MAX,
     UINT32_MAX,
@@ -228,8 +231,61 @@ export const TlvPercent = TlvUInt8.bound({ max: 100 });
 export const TlvPercent100ths = TlvUInt16.bound({ max: 10000 });
 
 // Time Number types
-export const TlvEpochUs = TlvUInt64;
-export const TlvEpochS = TlvUInt32;
 export const TlvPosixMs = TlvUInt64;
 export const TlvSysTimeUs = TlvUInt64;
 export const TlvSysTimeMS = TlvUInt64;
+
+/** Milliseconds from Unix epoch (1970-01-01) to Matter epoch (2000-01-01) */
+export const MATTER_EPOCH_OFFSET = Days(10_957);
+
+/** Seconds from Unix epoch (1970-01-01) to Matter epoch (2000-01-01) */
+export const MATTER_EPOCH_OFFSET_S = Seconds.of(MATTER_EPOCH_OFFSET);
+
+/** Microseconds from Unix epoch (1970-01-01) to Matter epoch (2000-01-01) */
+export const MATTER_EPOCH_OFFSET_US = BigInt(MATTER_EPOCH_OFFSET * 1_000);
+
+/**
+ * TLV Schema for Epoch time in seconds since Matter epoch (2000-01-01). You can just use the normal unix epoch
+ * time (since 1970-01-01) number and it will be converted automatically.
+ */
+export const TlvEpochS = new TlvWrapper<number, number>(
+    new TlvNumberSchema(
+        TlvType.UnsignedInt,
+        value => TlvCodec.getUIntTlvLength(value),
+        0, // too low values will be caught in the wrapper
+        UINT32_MAX + MATTER_EPOCH_OFFSET_S,
+    ),
+    unixEpoch => {
+        const value = unixEpoch - MATTER_EPOCH_OFFSET_S;
+        if (value < 0) {
+            throw new ImplementationError(
+                "Do not convert Epoch-values yourself, use TlvEpochS directly with unix epoch values.",
+            );
+        }
+        return value;
+    },
+    epochS => epochS + MATTER_EPOCH_OFFSET_S,
+);
+
+/**
+ * TLV Schema for Epoch time in microseconds since Matter epoch (2000-01-01). You can just use the unix epoch as
+ * microseconds (since 1970-01-01) number and it will be converted automatically.
+ */
+export const TlvEpochUs = new TlvWrapper<number | bigint, number | bigint>(
+    new TlvLongNumberSchema(
+        TlvType.UnsignedInt,
+        value => TlvCodec.getUIntTlvLength(value),
+        0, // too low values will be caught in the wrapper
+        UINT64_MAX + MATTER_EPOCH_OFFSET_US,
+    ),
+    unixEpoch => {
+        const result = BigInt(unixEpoch) - MATTER_EPOCH_OFFSET_US;
+        if (result < BigInt(0)) {
+            throw new ImplementationError(
+                "Do not convert Epoch-values yourself, use TlvEpochUs directly with unix epoch values.",
+            );
+        }
+        return result;
+    },
+    epochUs => BigInt(epochUs) + MATTER_EPOCH_OFFSET_US,
+);